Array fixes and refactor tests
