+++ nullaway/src/main/java/com/uber/nullaway/NullabilityUtil.java
@@ -22,6 +22,9 @@
 
 package com.uber.nullaway;
 
+import static com.sun.tools.javac.code.TypeAnnotationPosition.TypePathEntryKind.ARRAY;
+import static com.sun.tools.javac.code.TypeAnnotationPosition.TypePathEntryKind.INNER_TYPE;
+
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableSet;
 import com.google.errorprone.VisitorState;
@@ -41,7 +44,7 @@ import com.sun.tools.javac.code.Attribute;
 import com.sun.tools.javac.code.Symbol;
 import com.sun.tools.javac.code.TargetType;
 import com.sun.tools.javac.code.Type;
-import com.sun.tools.javac.code.TypeAnnotationPosition;
+import com.sun.tools.javac.code.TypeAnnotationPosition.TypePathEntry;
 import com.sun.tools.javac.code.Types;
 import com.sun.tools.javac.tree.JCTree;
 import com.sun.tools.javac.util.JCDiagnostic;
@@ -293,10 +296,35 @@ public class NullabilityUtil {
     // location is a list of TypePathEntry objects, indicating whether the annotation is
     // on an array, inner type, wildcard, or type argument. If it's empty, then the
     // annotation is directly on the type.
-    // We care about both annotations directly on the outer type and also those on directly
-    // on an inner type, but wish to discard annotations on arrays, wildcards, or type arguments.
-    return t.position.location.stream()
-        .allMatch(l -> l.tag.equals(TypeAnnotationPosition.TypePathEntryKind.INNER_TYPE));
+    // We care about both annotations directly on the outer type and also those directly
+    // on an inner type or array dimension, but wish to discard annotations on wildcards,
+    // or type arguments.
+    // For arrays, we treat annotations on the outer type and on any dimension of the array
+    // as applying to the nullability of the array itself, not the elements.
+    // We don't allow mixing of inner types and array dimensions in the same location
+    // (i.e. `Foo.@Nullable Bar []` is meaningless).
+    // These aren't correct semantics for type use annotations, but a series of hacky
+    // compromises to keep some semblance of backwards compatibility until we can do a
+    // proper deprecation of the incorrect behaviors for type use annotations when their
+    // semantics don't match those of a declaration annotation in the same position.
+    // See https://github.com/uber/NullAway/issues/708
+    boolean location_has_inner_types = false;
+    boolean location_has_array = false;
+    for (TypePathEntry entry : t.position.location) {
+      switch (entry.tag) {
+        case INNER_TYPE:
+          location_has_inner_types = true;
+          break;
+        case ARRAY:
+          location_has_array = true;
+          break;
+        default:
+          // Wildcard or type argument!
+          return false;
+      }
+    }
+    // Make sure it's not a mix of inner types and arrays for this annotation's location
+    return !(location_has_inner_types && location_has_array);
   }
 
   /**
