+++ nullaway/src/test/java/com/uber/nullaway/NullAwayCoreTests.java
@@ -960,164 +960,4 @@ public class NullAwayCoreTests extends NullAwayTestsBase {
             "}")
         .doTest();
   }
-
-  @Test
-  public void annotationAppliedToTypeParameter() {
-    defaultCompilationHelper
-        .addSourceLines(
-            "Test.java",
-            "package com.uber;",
-            "import java.util.List;",
-            "import java.util.ArrayList;",
-            "import org.checkerframework.checker.nullness.qual.Nullable;",
-            "class TypeArgumentAnnotation {",
-            "  List<@Nullable String> fSafe = new ArrayList<>();",
-            "  @Nullable List<String> fUnsafe = new ArrayList<>();",
-            "  void useParamSafe(List<@Nullable String> list) {",
-            "    list.hashCode();",
-            "  }",
-            "  void useParamUnsafe(@Nullable List<String> list) {",
-            "    // BUG: Diagnostic contains: dereferenced",
-            "    list.hashCode();",
-            "  }",
-            "  void useFieldSafe() {",
-            "    fSafe.hashCode();",
-            "  }",
-            "  void useFieldUnsafe() {",
-            "    // BUG: Diagnostic contains: dereferenced",
-            "    fUnsafe.hashCode();",
-            "  }",
-            "}")
-        .doTest();
-  }
-
-  @Test
-  public void annotationAppliedToInnerTypeImplicitly() {
-    defaultCompilationHelper
-        .addSourceLines(
-            "Test.java",
-            "package com.uber;",
-            "import org.checkerframework.checker.nullness.qual.Nullable;",
-            "class Test {",
-            "  @Nullable Foo f;", // i.e. Test.@Nullable Foo
-            "  class Foo { }",
-            "  public void test() {",
-            "    // BUG: Diagnostic contains: dereferenced",
-            "    f.hashCode();",
-            "  }",
-            "}")
-        .doTest();
-  }
-
-  @Test
-  public void annotationAppliedToInnerTypeExplicitly() {
-    defaultCompilationHelper
-        .addSourceLines(
-            "Test.java",
-            "package com.uber;",
-            "import org.checkerframework.checker.nullness.qual.Nullable;",
-            "class Test {",
-            "  Test.@Nullable Foo f1;",
-            "  @Nullable Test.Foo f2;",
-            "  class Foo { }",
-            "  public void test() {",
-            "    // BUG: Diagnostic contains: dereferenced",
-            "    f1.hashCode();",
-            "    // BUG: Diagnostic contains: dereferenced",
-            "    f2.hashCode();",
-            "  }",
-            "}")
-        .doTest();
-  }
-
-  @Test
-  public void annotationAppliedToInnerTypeExplicitly2() {
-    defaultCompilationHelper
-        .addSourceLines(
-            "Bar.java",
-            "package com.uber;",
-            "import org.checkerframework.checker.nullness.qual.Nullable;",
-            "class Bar {",
-            "  public class Foo { }",
-            "}")
-        .addSourceLines(
-            "Test.java",
-            "package com.uber;",
-            "import org.checkerframework.checker.nullness.qual.Nullable;",
-            "class Test {",
-            "  Bar.@Nullable Foo f1;",
-            "  @Nullable Bar.Foo f2;",
-            "  public void test() {",
-            "    // BUG: Diagnostic contains: dereferenced",
-            "    f1.hashCode();",
-            "    // BUG: Diagnostic contains: dereferenced",
-            "    f2.hashCode();",
-            "  }",
-            "}")
-        .doTest();
-  }
-
-  @Test
-  public void annotationAppliedToInnerTypeOfTypeArgument() {
-    defaultCompilationHelper
-        .addSourceLines(
-            "Test.java",
-            "package com.uber;",
-            "import java.util.Set;",
-            "import org.checkerframework.checker.nullness.qual.Nullable;",
-            "class Test {",
-            "  // BUG: Diagnostic contains: @NonNull field s not initialized",
-            "  Set<@Nullable Foo> s;", // i.e. Set<Test.@Nullable Foo>
-            "  class Foo { }",
-            "  public void test() {",
-            "    // safe because field is @NonNull",
-            "    s.hashCode();",
-            "  }",
-            "}")
-        .doTest();
-  }
-
-  @Test
-  public void typeUseAnnotationOnArray() {
-    defaultCompilationHelper
-        .addSourceLines(
-            "Test.java",
-            "package com.uber;",
-            "import org.checkerframework.checker.nullness.qual.Nullable;",
-            "class Test {",
-            "  // ok only for backwards compat",
-            "  @Nullable Object[] foo1 = null;",
-            "  // ok according to spec",
-            "  Object @Nullable[] foo2 = null;",
-            "  // ok only for backwards compat",
-            "  @Nullable Object [][] foo3 = null;",
-            "  // ok according to spec",
-            "  Object @Nullable [][] foo4 = null;",
-            "  // NOT ok; @Nullable applies to first array dimension",
-            "  // BUG: Diagnostic contains: assigning @Nullable expression",
-            "  Object [] @Nullable [] foo5 = null;",
-            "}")
-        .doTest();
-  }
-
-  @Test
-  public void typeUseAnnotationOnInnerMultiLevel() {
-    defaultCompilationHelper
-        .addSourceLines(
-            "Test.java",
-            "package com.uber;",
-            "import java.util.Set;",
-            "import org.checkerframework.checker.nullness.qual.Nullable;",
-            "class A { class B { class C {} } }",
-            "class Test {",
-            "  // At some point, we should not treat the foo1 and foo2 as @Nullable.",
-            "  // For now we do, for ease of compatibility.",
-            "  @Nullable A.B.C foo1 = null;",
-            "  A.@Nullable B.C foo2 = null;",
-            "  A.B.@Nullable C foo3 = null;",
-            "  // what should we do here?",
-            "  A.B.@Nullable C [][] foo4 = null;",
-            "}")
-        .doTest();
-  }
 }
